refactor: rename app2.py to routes.py
