+++ tests/test_routes.py
@@ -1,6 +1,6 @@
 import pytest
 
-from sketch_map_tool.app2 import app
+from sketch_map_tool.routes import app
 
 
 @pytest.fixture()
